ircd::json: Remove unit linkages for constant expression values.

+++ include/ircd/json/iov.h
@@ -55,10 +55,7 @@ struct ircd::json::iov
 	IRCD_EXCEPTION(error, exists);
 	IRCD_EXCEPTION(error, oversize);
 
-	static constexpr const size_t &max_size
-	{
-		1024
-	};
+	static constexpr const uint &max_size {1024};
 
   public:
 	bool has(const string_view &key) const;
